Implement filter for saved assets
The filter function runs in O(n) time | O(n) space where n = total
number of assets returned by the API. The filter method iterates over n
elements but retrieving the asset id and checking if that id is in the
hash-table are constant time operations.

The current linear time complexity could be improved to O(log n) if we
employed a binary search or a quick search but that would require a
sorted array by IDs. Whilst both search methods are done in place the
space complexity for a filter would not change as we would still need
to create a new sub-array. I suspect that the size of n is unlikely to
reach a level at Atmos where the additional overhead of ensuring order
is a worthwhile trade-off. Hence, I will use the linear implementation.

Some flexbox styling changes are made in asset.scss and
showInventory.scss.