Implement modal feature
If a user clicks on an asset component, the click handler executes a
function that injects a query string of this signature into the url:
?[homes | lets]=id

A useEffect checks for changes in the queryString object and if the
query string's key and value match with the assetType and assetId, the
modal is opened.

The modal features a button that allows the user to close the modal
again.

Some changes to the Readme were made and some styling was added to
the modal.