Implement Modal view of compatible assets
Modal displays any assets that are compatible with the current asset.
However, the favourites button does not work and clicking on the asset
component wants to inject a new query string into the url. This is
because currently the asset component include features that are not
relevant for this representation. A better way can probably be found
to cross-cut concerns by using custom hooks or HOCs.