Implement HOC withModal feature
The asset component can now be rendered without a modal or with one. The
latter has been implemented using a withModal HOC. The coupling is still
not as loose as I would like it to be.

Also, the url with a query string can now be shared with other users,
who will then be able to see the modal of the asset and its
complementary assets.